4.21 step9 done

+++ controllers/blogs.js
@@ -61,8 +61,22 @@ blogsRouter.put('/:id', async (request, response) => {
 });
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
+  const blogToDelete = await Blog.findById(request.params.id);
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  console.log('decoded token: ', decodedToken);
+
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' });
+  }
+
+  if (blogToDelete.user.id.toString() === decodedToken.id) {
+    await Blog.findByIdAndRemove(request.params.id);
+    response.status(204).end();
+  }
+  else {
+    return response.status(401).json( { error: 'You are not allowed to remove this blog' });
+  }
 });
 
 module.exports = blogsRouter;
